Add files via upload
In this Employeelist.csv file details of employees has been mentioned. I Have created the few PHP pages for the specific feature. Based on the required feature one can retrive it directly. I need to optimize it. User can get the file from the Bio-metric machine and they need to have some sort of CSV file in which emoloyee details should be linked with the file in the bio-metric file.